+++ src/App.tsx
@@ -10,7 +10,8 @@ import { ThemeProvider } from "@/hooks/use-theme";
 import { Toaster } from "@/components/ui/toaster";
 import { Navigation } from "@/components/Navigation";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { AuthProvider, useAuth } from "@/contexts/AuthContext";
+import { AuthProvider } from "@/contexts/AuthContext";
+import { useAuth } from "@/hooks/useAuth";
 import { Suspense, lazy, useEffect } from "react";
 import { SpeedInsights } from "@vercel/speed-insights/react";
 import { Analytics, track } from "@vercel/analytics/react";
