+++ src/pages/Index.tsx
@@ -9,7 +9,7 @@ import {
   BrainIcon,
   UsersIcon,
 } from "lucide-react";
-import { useAuth } from "@/contexts/AuthContext";
+import { useAuth } from "@/hooks/useAuth";
 
 const Index = () => {
   const navigate = useNavigate();
